% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{blg_bind_list_rows}
\alias{blg_bind_list_rows}
\title{Bind a list of 'rows' (vectors) into a single tibble.}
\usage{
blg_bind_list_rows(list_rows)
}
\arguments{
\item{list_rows}{A list containing 'rows' of data (vectors).}
}
\description{
This is a wrapper of `data.table::rbindlist()` that
converts the final output to a tibble.
}
