% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{blg_bind_list_tibbles}
\alias{blg_bind_list_tibbles}
\title{Combine (union) a list of data frames or tibbles into a
single tibble.}
\usage{
blg_bind_list_tibbles(list_tibbles)
}
\arguments{
\item{list_tibbles}{A list containing tibbles or data.frames.}
}
\description{
Should just use `dplyr::bind_rows()`
}
