% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{blg_days_in_year}
\alias{blg_days_in_year}
\title{Find the number of days in a date's year.}
\usage{
blg_days_in_year(date)
}
\arguments{
\item{date}{A date object.}
}
\description{
Enter a date and get the number of days in that date's year.
}
