% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_cap}
\alias{blg_frmt_cap}
\title{Capitalize the first word in a string, not the
individual words.}
\usage{
blg_frmt_cap(string)
}
\arguments{
\item{string}{A single character string.}
}
\description{
Capitalize the first word in a string, not the
individual words.
}
