% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_capitalize}
\alias{blg_frmt_capitalize}
\title{Capitalize the first word in a string, not the
individual words.}
\usage{
blg_frmt_capitalize()
}
\description{
Capitalize the first word in a string, not the
individual words.
}
