% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_erase}
\alias{blg_frmt_erase}
\title{Delete every other label in a character vector.}
\usage{
blg_frmt_erase(eliminate_which = c("even", "odd"))
}
\arguments{
\item{eliminate_which}{String indicating whether to erase the even (default)
or odd labels.}
}
\description{
Assign blank `""` (erase) values to every other string in a character vector.
Convenient for eliminating every other label in a plot without removing the
grid lines.
}
