% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_percent}
\alias{blg_frmt_percent}
\title{Control how percentages are printed.}
\usage{
blg_frmt_percent(x, i = 1)
}
\arguments{
\item{x}{A number (percent) to be formatted.}

\item{i}{Number of figures to the right
of the decimal to show (default = 1)}
}
\description{
Uses `sprintf()` to truncate decimal and print as a percent.
}
