% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_stagger}
\alias{blg_frmt_stagger}
\title{Add a prefixed line break literal to every other string in a character
vector.}
\usage{
blg_frmt_stagger(n = 1)
}
\arguments{
\item{n}{Number of line breaks to prefix to each string.}
}
\description{
Prefix every other string in a character vector with a line break. Number of
breaks determined by `n`. Useful for staggering overlapping labels on an
x-axis.
}
