% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_wrap}
\alias{blg_frmt_wrap}
\title{Wrap long strings with a line break literal}
\usage{
blg_frmt_wrap(length = 18)
}
\arguments{
\item{length}{A number specifying how many characters to include in each line
before inserting the line break literal.}
}
\description{
Wraps `stringi::stri_wrap()` with a literal line break 'prefix' and a blank
'initial'. Ment for use with 'ggplot2'.
}
