% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{blg_frmt_wrap_stagger}
\alias{blg_frmt_wrap_stagger}
\title{Wrap and stagger labels.}
\usage{
blg_frmt_wrap_stagger(length = 18, n = 1)
}
\arguments{
\item{length}{A number specifying how many characters to include in each line
before inserting the line break literal.}

\item{n}{Number of line breaks to prefix to each string.}
}
\description{
Applies a wrapping and staggering line break.
}
