% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{blg_source_value}
\alias{blg_source_value}
\title{Collect only the 'value' from the `source()` function.}
\usage{
blg_source_value(file)
}
\arguments{
\item{file}{A `.R` file to source into the environment.}
}
\description{
An explicit and easy-to-understand method of extracting
the value from the `source()` function.
}
