% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{blg_theme_default}
\alias{blg_theme_default}
\title{A 'theme_minimal' based theme.}
\usage{
blg_theme_default(text_size = 16)
}
\arguments{
\item{text_size}{Base font size.}
}
\description{
This is 'theme_minimal' without minor grids and with
diferent text settings (Open Sans; 16).
}
