% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{blg_tibblefy}
\alias{blg_tibblefy}
\title{Convert parsed JSON data (list) into a nested tibble.}
\usage{
blg_tibblefy(x)
}
\arguments{
\item{x}{A list of vectors and data.frames (from parsed JSON).}
}
\description{
Converts vectors into columns and converts data frames
into list columns. Tibblefies data.drames and produces
a tibble whose columns represent the top most elements
of the list.
}
