% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_path}
\alias{file_path}
\title{A file path generator that ends with file separator.}
\usage{
file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{Characters to generate the file path}

\item{fsep}{The separator to use between characters and at the end of the
file path}
}
\description{
Generate file paths that end with the system's file separator, or whichever
other separator suits your needs
}
