% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_lines}
\alias{plot_lines}
\title{A new way of getting line plots of all the variables.}
\usage{
plot_lines(data, dc_order = 1L)
}
\arguments{
\item{data}{The data to plot}

\item{dc_order}{Which date column should be used as the x-axis. Default is the first ordinal occurring date column.}
}
\description{
Plot all numeric variables against any existing base or zoo date variable.
You can specify which by using the `dc_order` argument.
}
