% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_lookup}
\alias{state_lookup}
\title{State lookup tibble.}
\format{A US state lookup table of names and abbreviations}
\usage{
state_lookup
}
\description{
This sure seems useful when you need to match a state's abbreviation with its
full name, and vice-versa
}
\keyword{datasets}
